Sorta fixed schedule proposal page
