--- planner-frontend/src/pages/ShiftAvailability.css
@@ -1,74 +1,0 @@
-.shift-availability-container {
-    width: 100%;
-    padding: 20px;
-    font-family: Arial, sans-serif;
-}
-
-.name-input {
-    margin-bottom: 10px;
-}
-
-.name-input label {
-    font-weight: bold;
-    margin-right: 10px;
-}
-
-.name-input input {
-    padding: 5px;
-    width: 200px;
-}
-
-.shift-table {
-    width: 100%;
-    border-collapse: collapse;
-    margin-top: 20px;
-}
-
-.shift-table th, .shift-table td {
-    border: 1px solid #ddd;
-    padding: 8px;
-    text-align: center;
-}
-
-.shift-table th {
-    background-color: #007bff;
-    color: white;
-}
-
-.shift-preferences {
-    margin-top: 20px;
-    display: flex;
-    align-items: center;
-    gap: 10px;
-}
-
-.shift-preferences label {
-    font-weight: bold;
-}
-
-.shift-preferences input {
-    width: 60px;
-    padding: 5px;
-}
-
-.shift-summary {
-    margin-top: 10px;
-    font-size: 16px;
-}
-
-.error {
-    color: red;
-    font-weight: bold;
-}
-.shift-table td {
-    width: 60px;
-    height: 30px;
-    border: 1px solid #ddd;
-    cursor: pointer;
-  }
-  
-  .shift-table td.selected {
-    background-color: #4caf50;
-    color: white;
-  }
-
