+++ planner-frontend/src/pages/styling/ShiftAvailability.css
@@ -1,0 +1,102 @@
+.shift-availability-container {
+    display: flex;
+    justify-content: center;
+    height: 100vh;
+    overflow: hidden;
+}
+
+.sidebar-container {
+    width: 250px;
+    background-color: #2c3e50;
+    color: white;
+    padding: 20px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+}
+
+/* Table styling */
+.shift-table {
+    width: 100%;
+    border-collapse: collapse;
+    margin-top: 20px;
+}
+
+.shift-table th, .shift-table td {
+    border: 1px solid #ddd;
+    padding: 8px;
+    text-align: center;
+}
+
+.shift-table th {
+    background-color: #007bff;
+    color: white;
+}
+
+.shift-preferences {
+    margin-top: 20px;
+    display: flex;
+    align-items: center;
+    gap: 10px;
+}
+
+.shift-preferences label {
+    font-weight: bold;
+}
+
+.shift-preferences input {
+    width: 60px;
+    padding: 5px;
+}
+
+.shift-summary {
+    margin-top: 10px;
+    font-size: 16px;
+}
+
+.error {
+    color: red;
+    font-weight: bold;
+}
+
+.shift-table td {
+  width: 60px;
+  height: 30px;
+  border: 1px solid #ddd;
+  cursor: pointer;
+}
+
+.shift-table td.selected {
+  background-color: #28a745;
+  color: white;
+}
+
+    /* Button Styling */
+.submit-button {
+    margin-top: 20px;
+    padding: 10px 20px;
+    font-size: 16px;
+    color: white;
+    background-color: #007bff;
+    border: none;
+    cursor: pointer;
+    border-radius: 5px;
+    display: block;
+    margin-left: auto;
+    margin-right: auto;
+}
+
+.submit-button:hover {
+  background-color: #0056b3;
+}
+
+.proposal-sub-container button {
+    width: 100%;
+    padding: 10px;
+    margin-top: 10px;
+    background-color: #007bff;
+    color: #fff;
+    border: none;
+    border-radius: 4px;
+    cursor: pointer;
+  }
